feat(babylon): WorldCreator introduced locking logic, moved gizmo stuff
