+++ packages/odyssey3d/src/babylon/ObjectHelper.ts
@@ -6,7 +6,6 @@ import {
   Engine,
   AssetContainer,
   Matrix,
-  GizmoManager,
   Mesh,
   InstantiatedEntries,
   Behavior,
@@ -21,6 +20,7 @@ import {Object3dInterface, Texture3dInterface} from '@momentum-xyz/core';
 
 import {CameraHelper} from './CameraHelper';
 import {SkyboxHelper} from './SkyboxHelper';
+import {WorldCreatorHelper} from './WorldCreatorHelper';
 
 interface BabylonObjectInterface {
   container: AssetContainer;
@@ -32,19 +32,11 @@ class CustomNode extends TransformNode {
   public onSomeChange = new Observable();
 }
 
-export enum GizmoTypesEnum {
-  Position,
-  Rotation,
-  Scale,
-  BoundingBox
-}
-
 export class ObjectHelper {
   static light: HemisphericLight | null = null;
   static assetRootUrl = 'https://dev2.odyssey.ninja/api/v3/render/asset/';
   static textureRootUrl = 'https://dev2.odyssey.ninja/api/v3/render/texture/';
   static textureDefaultSize = 's3/';
-  static gizmoManager: GizmoManager;
   static objectsMap = new Map<string, BabylonObjectInterface>();
   static followObjectBehaviour: Behavior<InstantiatedEntries>;
   static player: TransformNode;
@@ -81,17 +73,16 @@ export class ObjectHelper {
           while (parent.parent) {
             parent = parent.parent as AbstractMesh;
           }
+          if (WorldCreatorHelper.isCreatorMode) {
+            WorldCreatorHelper.tryLockObject(parent.metadata);
+          }
+          // feclient.sendclick(parent.metadata);
           console.log(parent.metadata);
+        } else {
+          WorldCreatorHelper.unlockLastObject();
         }
       }
     };
-
-    // Gizmo setup
-    this.gizmoManager = new GizmoManager(scene);
-    this.gizmoManager.clearGizmoOnEmptyPointerEvent = true;
-
-    // Enable position gizmo by default
-    this.setGizmoType(GizmoTypesEnum.Position);
   }
 
   static setWorld(assetID: string) {
@@ -184,7 +175,7 @@ export class ObjectHelper {
     this.objectsMap.set(object.id, babylonObject);
 
     // Attach gizmo to object when clicked
-    this.gizmoManager.attachableNodes?.push(node);
+    WorldCreatorHelper.gizmoManager.attachableNodes?.push(node);
 
     // Play animations
     for (const group of instance.animationGroups) {
@@ -210,36 +201,6 @@ export class ObjectHelper {
       this.objectsMap.delete(id);
     } else {
       console.log("unable to delete object, as the id doesn't exist in the map, " + id);
-    }
-  }
-
-  static setGizmoType(type: GizmoTypesEnum) {
-    this.disableAllGizmos();
-    switch (type) {
-      case GizmoTypesEnum.Position:
-        this.gizmoManager.positionGizmoEnabled = true;
-        break;
-      case GizmoTypesEnum.Rotation:
-        this.gizmoManager.rotationGizmoEnabled = true;
-        break;
-      case GizmoTypesEnum.Scale:
-        this.gizmoManager.scaleGizmoEnabled = true;
-        break;
-      case GizmoTypesEnum.BoundingBox:
-        this.gizmoManager.boundingBoxGizmoEnabled = true;
-        break;
-
-      default:
-        break;
-    }
-  }
-
-  static disableAllGizmos() {
-    if (this.gizmoManager) {
-      this.gizmoManager.positionGizmoEnabled = false;
-      this.gizmoManager.rotationGizmoEnabled = false;
-      this.gizmoManager.scaleGizmoEnabled = false;
-      this.gizmoManager.boundingBoxGizmoEnabled = false;
     }
   }
